executor: migrate test-infra to testify for executor/write_test.go (#29953)
