+++ executor/aggregate_test.go
@@ -651,15 +651,6 @@ func TestGroupConcatAggr(t *testing.T) {
 	rows.Check(testkit.Rows("01234567", "12345"))
 }
 
-func fillData(tk *testkit.TestKit, table string) {
-	tk.MustExec("use test")
-	tk.MustExec(fmt.Sprintf("create table %s(id int not null default 1, name varchar(255), PRIMARY KEY(id));", table))
-
-	// insert data
-	tk.MustExec(fmt.Sprintf("insert INTO %s VALUES (1, \"hello\");", table))
-	tk.MustExec(fmt.Sprintf("insert into %s values (2, \"hello\");", table))
-}
-
 func TestSelectDistinct(t *testing.T) {
 	t.Parallel()
 	store, clean := testkit.CreateMockStore(t)
