+++ testkit/testkit.go
@@ -228,6 +228,12 @@ func (tk *TestKit) MustUseIndex(sql string, index string, args ...interface{}) b
 	return false
 }
 
+// CheckExecResult checks the affected rows and the insert id after executing MustExec.
+func (tk *TestKit) CheckExecResult(affectedRows, insertID int64) {
+	tk.require.Equal(int64(tk.Session().AffectedRows()), affectedRows)
+	tk.require.Equal(int64(tk.Session().LastInsertID()), insertID)
+}
+
 // WithPruneMode run test case under prune mode.
 func WithPruneMode(tk *TestKit, mode variable.PartitionPruneMode, f func()) {
 	tk.MustExec("set @@tidb_partition_prune_mode=`" + string(mode) + "`")
